+++ sdf_d3d12/src/Renderer/Buffer/D3DUploadBuffer.h
@@ -47,7 +47,7 @@ public:
 		if (alignment > 0)
 		{
 			// Align m_ElementStride to 256 bytes
-			m_ElementStride = Align(m_ElementStride, m_Alignment);
+			m_ElementStride = static_cast<UINT>(static_cast<UINT>(Align(m_ElementStride, m_Alignment)));
 		}
 
 		m_InstanceStride = m_ElementStride * m_ElementCount;
